+++ application/modules/forum/views/topic.php
@@ -29,7 +29,7 @@ endif; ?>
               </div>
               <div class="uk-width-auto">
                 <?php if($this->wowauth->isLogged()): ?>
-                <?php if($this->forum_model->getSpecifyPostAuthor($idlink) == $this->session->userdata('wow_sess_id')): ?>
+                <?php if($this->forum_model->getSpecifyPostAuthor($idlink) == $this->session->userdata('id')): ?>
                 <div class="uk-text-center uk-text-right@s">
                   <a href="#" class="uk-button uk-button-default uk-button-small" uk-toggle="target: #editTopic"><i class="far fa-edit"></i> <?= lang('button_edit_topic'); ?></a>
                 </div>
@@ -87,7 +87,7 @@ endif; ?>
                 <div class="uk-width-expand@s">
                   <p class="uk-text-small uk-text-meta uk-margin-remove"><?= date('F d Y - H:i A', $commentss->date); ?></p>
                   <?= $commentss->commentary ?>
-                  <?php if($this->wowauth->getRank($this->session->userdata('wow_sess_id')) > 0 || $this->session->userdata('wow_sess_id') == $commentss->author && now() < strtotime('+30 minutes', $commentss->date)): ?>
+                  <?php if($this->wowauth->getRank($this->session->userdata('id')) > 0 || $this->session->userdata('id') == $commentss->author && now() < strtotime('+30 minutes', $commentss->date)): ?>
                   <div class="uk-margin-small-top uk-margin-remove-bottom">
                     <button class="uk-button uk-button-danger uk-button-small" value="<?= $commentss->id ?>" id="button_delete<?= $commentss->id ?>" onclick="DeleteTopicReply(event, this.value)"><i class="fas fa-eraser"></i> <?= lang('button_remove'); ?></button>
                   </div>