+++ application/modules/store/views/cart.php
@@ -28,7 +28,7 @@
                       <div class="uk-form-controls uk-light">
                         <select class="uk-select uk-width-1-1" onchange="updateCharacter(this, '<?php echo $item["rowid"]; ?>')">
                           <option value="0"><?= lang('notification_select_character'); ?></option>
-                          <?php foreach($this->wowrealm->getGeneralCharactersSpecifyAcc($this->wowrealm->getRealmConnectionData($this->store_model->getCategoryRealmId($item["category"])) ,$this->session->userdata('wow_sess_id'))->result() as $listchar): ?>
+                          <?php foreach($this->wowrealm->getGeneralCharactersSpecifyAcc($this->wowrealm->getRealmConnectionData($this->store_model->getCategoryRealmId($item["category"])) ,$this->session->userdata('id'))->result() as $listchar): ?>
                           <option value="<?= $listchar->guid ?>" <?php if($listchar->guid == $item["guid"]) echo 'selected'; ?>><?= $listchar->name ?> - (<?= lang('table_header_realm'); ?>: <?= $this->wowrealm->getRealmName($this->store_model->getCategoryRealmId($item["category"])); ?>)</option>
                           <?php endforeach; ?>
                         </select>